requirements_sphere.txt
absl-py==2.0.0
addict==2.4.0
ansi2html==1.8.0
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
cachetools==5.3.2
ccimport==0.4.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coalesce==0.3
comm==0.1.4
ConfigArgParse==1.7
cumm-cu114==0.2.8
cycler==0.11.0
dash==2.14.1
nuscenes-devkit
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
easydict==1.11
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.18.1
filelock==3.12.2
fire==0.5.0
Flask==2.2.5
fonttools==4.38.0
fsspec==2023.1.0
gmplot==1.4.1
grpcio==1.59.2
idna==3.4
imageio==2.31.2
importlib-metadata==4.13.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lark==1.1.8
llvmlite==0.39.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
msgpack==1.0.5
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.7.0
nest-asyncio==1.5.8
networkx==2.6.3
ninja==1.11.1.1
notebook==6.5.6
notebook_shim==0.2.4
numba==0.56.4
numpy==1.21.6
oauthlib==3.2.2
open3d==0.17.0
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.3
pccm==0.4.9
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.40
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-louvain==0.16
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
ray==1.10.0
rdflib==6.3.2
redis==5.0.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
safetensors==0.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.2
SharedArray==3.2.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
spconv-cu114==2.1.21
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.10
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transforms3d==0.4.1
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.9
zipp==3.15.0
